Fixed RC1 code issues in ReadMe.
Geolocation.Default and Map.Default didn't work, you'll get a build error because .Default isn't a known member. .Current will work.

Copying out "GetClosestAsync" and then replacing it right away with "GetClosestMonkey" seemed like it could be confusing. if someone just copied the stub of the method and then the method body from the "GetClosestMonkey" sample, they'd get a build error after adding the UI code, because a command with an incorrect name would be generate.